Potential 64 bit compatitibility bug fixed
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@658 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Sim/Units/COB/CobFile.cpp
@@ -140,9 +140,9 @@ CCobFile::CCobFile(CFileHandler &in, string name)
 
 	// If this is a TA:K script, read the sound names
 	if (ch.VersionSignature == 6) {
-		for (long i = 0; i < ch.NumberOfSounds; ++i) {
-			long ofs;
-			ofs = *(long *)&cobdata[ch.OffsetToSoundNameArray + i * 4];
+		for (int i = 0; i < ch.NumberOfSounds; ++i) {
+			int ofs;
+			ofs = *(int *)&cobdata[ch.OffsetToSoundNameArray + i * 4];
 			/* TODO This probably isn't correct. */
 			ofs = swabdword(ofs);
 			string s = &cobdata[ofs];
